+++ trunk/sfox/src/starengine/engine/heightfield.h
@@ -9,11 +9,11 @@ typedef struct heightfield *heightfield;
 /*image size must be (2^n)+1 x (2^m)+1*/
 extern heightfield heightfield_create_from_file(camera cam,
                                                 char *heightmap_filename,
-                                                double sizex, double sizey,
-                                                double sizez);
+                                                float sizex, float sizey,
+                                                float sizez);
 
 extern void heightfield_set_textures_from_file(heightfield hf, char *land, char *details);
-extern void heightfield_set_detail_scale(heightfield hf, double sx, double sy);
+extern void heightfield_set_detail_scale(heightfield hf, float sx, float sy);
 
 extern void heightfield_to_opengl(heightfield hf);
