+++ trunk/sfox/src/starengine/engine/heightfieldinfo.h
@@ -7,7 +7,7 @@
 
 typedef struct heightfieldinfo *heightfieldinfo;
 
-extern heightfieldinfo heightfieldinfo_create(char *landmap, heightfield hf, camera cam, display disp, double sizex, double sizey);
+extern heightfieldinfo heightfieldinfo_create(char *landmap, heightfield hf, camera cam, display disp, float sizex, float sizey);
 extern void heightfieldinfo_to_opengl(heightfieldinfo hfi);
 extern void heightfieldinfo_destroy(heightfieldinfo hfi);
